Remove interfaces with one implementation
